refactor(deps): Remove old AMQP websocket implementation
We use a published library for AMQPS_WS that we share with the event hubs team, so we don't need this local implementation
